feat: extract contract error mapping to contractErrors.ts
