+++ frontend/src/services/stellar.ts
@@ -36,35 +36,7 @@ function hexToBytes(hex: string): Uint8Array {
 }
 
 // ── Contract error codes ──────────────────────────────────────────────────────
-
-const CONTRACT_ERRORS: Record<number, string> = {
-  1: 'Insufficient fee payment. Please increase the fee amount.',
-  2: 'Unauthorized. You do not have permission to perform this action.',
-  3: 'Invalid parameters provided.',
-  4: 'Token not found.',
-  5: 'Metadata has already been set for this token.',
-  6: 'Contract is already initialized.',
-  7: 'Burn amount exceeds your token balance.',
-  8: 'Burning is not enabled for this token.',
-  9: 'Invalid burn amount. Must be greater than zero.',
-  10: 'Contract is paused. Please try again later.',
-}
-
-function parseContractError(err: unknown): Error {
-  const msg = err instanceof Error ? err.message : String(err)
-
-  // Soroban contract errors surface as "Error(Contract, X)" in the result XDR
-  const match = msg.match(/Error\(Contract,\s*(\d+)\)/)
-  if (match?.[1]) {
-    const code = parseInt(match[1], 10)
-    return new Error(CONTRACT_ERRORS[code] ?? `Contract error code ${code}`)
-  }
-
-  // Simulation failure messages
-  if (msg.includes('simulation')) return new Error(`Simulation failed: ${msg}`)
-
-  return err instanceof Error ? err : new Error(msg)
-}
+import { parseContractError } from '../utils/contractErrors'
 
 // ── Network helpers ───────────────────────────────────────────────────────────
